auth_service/pyproject.toml
[project]
name = "auth_service"
version = "0.1.0"
description = "Auth Service — аутентификация и управление пользователями BioFabric ERP"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "psycopg[binary]>=3.2.0",
  "pydantic>=2.9.0",
  "PyJWT>=2.9.0",
  "passlib[bcrypt]>=1.7.4",
  "bcrypt>=4.0.0",
]
